capabilities: intelligent system server cap drop
This patch cleans up the capability dropping code and makes it able to
deal with a system in which there are already caps dropped when this
code is run. New zygote processes are now started with capabilities
built from the intersection of the current caps and the desired caps.
Before the code always assumed that zygote is run with *all* caps
present.

Change-Id: Ib21012d0f487755ce5a2be0f154bca5911b004ba